+++ reductstore/src/api/entry/remove_batched.rs
@@ -72,7 +72,7 @@ mod tests {
         #[future] empty_body: Body,
     ) {
         headers.insert("content-length", "0".parse().unwrap());
-        headers.insert("x-reduct-time-xxx", "10".parse().unwrap());
+        headers.insert("x-reduct-time-yyy", "10".parse().unwrap());
 
         let err = remove_batched_records(
             State(components.await),
@@ -88,7 +88,7 @@ mod tests {
             err,
             HttpError::new(
                 ErrorCode::UnprocessableEntity,
-                "Invalid header'x-reduct-time-xxx': must be an unix timestamp in microseconds",
+                "Invalid header 'x-reduct-time-yyy': must be an unix timestamp in microseconds",
             )
         );
     }