+++ reductstore/src/api/entry/write_batched.rs
@@ -310,7 +310,7 @@ mod tests {
         #[future] body_stream: Body,
     ) {
         headers.insert("content-length", "10".parse().unwrap());
-        headers.insert("x-reduct-time-xxx", "10".parse().unwrap());
+        headers.insert("x-reduct-time-yyy", "10".parse().unwrap());
 
         let err = write_batched_records(
             State(components.await),
@@ -326,7 +326,7 @@ mod tests {
             err,
             HttpError::new(
                 ErrorCode::UnprocessableEntity,
-                "Invalid header'x-reduct-time-xxx': must be an unix timestamp in microseconds",
+                "Invalid header 'x-reduct-time-yyy': must be an unix timestamp in microseconds",
             )
         );
     }