+++ CHANGELOG.md
@@ -23,7 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- RS-446: Fix storage engine hanging during replication, [PR-564](https://github.com/reductstore/reductstore/pull/564)
+- RS-446: Storage engine hanging during replication, [PR-564](https://github.com/reductstore/reductstore/pull/564)
+- RS-468: Name of invalid batch header, [PR-586](https://github.com/reductstore/reductstore/pull/586)
 - Server shutdown, [PR-557](https://github.com/reductstore/reductstore/pull/557)
 - Internal 500 error after removing bucket or entry, [PR-565](https://github.com/reductstore/reductstore/pull/565)